python-agent/requirements.txt
google-auth-oauthlib>=1.2.0
google-auth>=2.23.0
google-api-python-client>=2.118.0
